+++ R/grec-main.R
@@ -58,9 +58,10 @@
 #'
 #' (*) Due to the deprecation of the \pkg{raster} package, \pkg{grec} will not be
 #' supporting the use of `RasterLayer` in future versions. Instead,
-#' \pkg{grec} will be incorporating support for [SpatRaster-class][terra::SpatRaster-class],
-#' a more recent and actively developed method for working with raster data.
-#' This change will take effect as soon as \pkg{raster} is removed from CRAN.
+#' we encourage the use of [SpatRaster-class][terra::SpatRaster] (from the
+#' \pkg{terra} package), a more recent and actively developed method for working
+#' with raster data. This change will take effect as soon as \pkg{raster} is
+#' removed from CRAN.
 #'
 #' Until the current version, `grec` performs four methods:
 #' \enumerate{
@@ -115,6 +116,12 @@
 #' as it is primarily intended for chlorophyll data. Users are free to apply the
 #' transformation manually if it suits their specific needs.
 #'
+#' @returns
+#' The output class will depend on the input (i.e. the \code{x} argument). For
+#' further details about structure of `SpatRaster` or `RasterLayer`, you can
+#' check the specific documentation in [terra][terra::SpatRaster] and
+#' [raster][raster::raster] respectively.
+#'
 #'
 #' @references
 #' Belkin, I. M., & O'Reilly, J. E. (2009). An algorithm for oceanic
@@ -126,8 +133,6 @@
 #' Pelagic Fish Species Using Catch Size- and Remote Sensing Data". Progress in
 #' Oceanography 59, No 2-3 (October 2003): 275-300.
 #' (\doi{https://doi.org/10.1016/j.pocean.2003.07.004}).
-#'
-#' @return The output will preserve the input class.
 #'
 #' @export
 #'
@@ -149,6 +154,8 @@
 #' # External transformation for chl data
 #' out_chl$z <- log10(out_chl$z)
 #'
+#' oldPar <- par(no.readonly = FALSE)
+#'
 #' par(mfrow = c(2, 2), mar = rep(0, 4), oma = rep(0, 4))
 #'
 #' image(exampleSSTData, col = colPalette, axes = FALSE)
@@ -163,6 +170,8 @@
 #' image(out_chl, col = colPalette, axes = FALSE)
 #' mtext(text = "Chlorophyll gradient\n(log scale)", side = 3, line = -4, adj = 0.99,
 #'       cex = 1.2)
+#'
+#' par(oldPar)
 getGradients <- function(x,
                          method = c("BelkinOReilly2009", "median_filter",
                                     "Agenbag2003-1", "Agenbag2003-2"),